+++ pkg/cmd/issue/view/view.go
@@ -8,8 +8,8 @@ import (
 	"strconv"
 
 	"github.com/spf13/cobra"
-	"github.com/qubernetic-org/copia-cli/pkg/cmdutil"
-	"github.com/qubernetic-org/copia-cli/pkg/iostreams"
+	"github.com/qubernetic/copia-cli/pkg/cmdutil"
+	"github.com/qubernetic/copia-cli/pkg/iostreams"
 )
 
 var validJSONFields = []string{"number", "title", "body", "state", "author", "labels", "created_at", "comments"}